refactor(minimina): restructure the supervisor into per-concern modules with backend traits
Replace the single-file supervisor with one concern per module:

- supervisor/plan.rs   — the input contract (node specs, backend spec,
  status). Plan builders and the runtime both depend on this module and
  neither on the other: the plan is the seam between "describe a network"
  and "run a network".
- supervisor/backend.rs — Backend/Unit/Killer traits: what a backend must
  provide (launch, wait, terminate/force-kill, teardown).
- supervisor/native.rs  — native impl (child processes, pdeathsig).
- supervisor/docker.rs  — docker impl (bollard containers + network).
- supervisor/rpc.rs     — JSON-RPC server + blocking client; request
  parsing hidden behind the RpcRequest type.
- supervisor/mod.rs     — the runtime: launch units with waiter tasks,
  serve RPC, tear down.

The RunningUnit/Killer enums become traits with associated types,
dispatched statically: run() matches BackendSpec exactly once and hands
off to the monomorphic run_backend<B: Backend>. A network is all native
or all docker; that single-axis invariant was previously a comment plus
runtime guards — now the type system enforces it (a NativeBackend can
only launch NativeNodeSpecs and yield native units) and the guards are
gone. Backend-level resources (the docker network + DNS) live on the
backend impl that owns them.

Per-node state also collapses: the parallel nodes/killers maps merge into
one HashMap<String, Node> where Node owns the status and, while the unit
is alive, its kill handle.

Behavior is unchanged; public surface (run_blocking, rpc_call, plan
types) is unchanged.
