--- src/app/minimina/src/supervisor.rs
@@ -1,879 +1,0 @@
-//! Foreground, tokio-managed network supervisor.
-//!
-//! `network start` builds a tokio runtime and calls [`run_blocking`]. The
-//! supervisor owns the network's **units** — native child processes *or* docker
-//! containers — reaping their real exit codes, and serves a hand-rolled
-//! **JSON-RPC 2.0** API over a Unix domain socket so a separate short-lived CLI
-//! invocation (`status`/`stop`) can drive it while it runs in the foreground.
-//!
-//! A "unit" is backend-agnostic: [`RunningUnit::Native`] wraps a
-//! `tokio::process::Child`; [`RunningUnit::Docker`] wraps a bollard container.
-//! Both expose the same spawn / wait-reap / kill shape, which is what lets the
-//! same supervisor own either backend (replacing `docker compose`).
-//!
-//! Scope today: no detachment (foreground); `status`/`stop` only; native path
-//! wired into the CLI, docker path exercised by an `#[ignore]`d integration test
-//! pending the `main.rs` docker plan builder (ticket `docker-bollard-supervisor`).
-//!
-//! Concurrency: one `Arc<Mutex<SupervisorState>>` (`std::sync::Mutex`, never held
-//! across `.await` — copy out, drop the guard, then await). Transport:
-//! newline-delimited JSON, interoperable with Go's `json.Encoder`/`json.Decoder`.
-
-use log::{error, info, warn};
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use std::path::{Path, PathBuf};
-use std::sync::{Arc, Mutex};
-
-use tokio::io::{AsyncBufReadExt, AsyncWriteExt, BufReader};
-use tokio::net::UnixListener;
-use tokio::sync::Notify;
-
-use bollard::container::{
-    Config, CreateContainerOptions, NetworkingConfig, RemoveContainerOptions, StopContainerOptions,
-    WaitContainerOptions,
-};
-use bollard::image::CreateImageOptions;
-use bollard::network::CreateNetworkOptions;
-use bollard::secret::{EndpointSettings, HostConfig, PortBinding};
-use bollard::Docker;
-use futures_util::StreamExt;
-
-// ---------------------------------------------------------------------------
-// Plan (subset of the materialized plan the supervisor needs)
-// ---------------------------------------------------------------------------
-
-/// A native daemon: a local process the supervisor spawns and owns.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NativeNodeSpec {
-    pub name: String,
-    pub binary: PathBuf,
-    pub args: Vec<String>,
-    #[serde(default)]
-    pub env: Vec<(String, String)>,
-    pub log_file: PathBuf,
-}
-
-/// A host↔container bind mount.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct Mount {
-    pub host: String,
-    pub container: String,
-    #[serde(default)]
-    pub read_only: bool,
-}
-
-/// A docker daemon: a container the supervisor creates, starts, and owns.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct DockerNodeSpec {
-    pub name: String,
-    pub image: String,
-    #[serde(default)]
-    pub entrypoint: Option<Vec<String>>,
-    #[serde(default)]
-    pub cmd: Vec<String>,
-    #[serde(default)]
-    pub env: Vec<(String, String)>,
-    /// (host_port, container_port) pairs to publish.
-    #[serde(default)]
-    pub ports: Vec<(u16, u16)>,
-    #[serde(default)]
-    pub mounts: Vec<Mount>,
-    /// Network aliases (service-name DNS) — replaces compose service names.
-    #[serde(default)]
-    pub aliases: Vec<String>,
-}
-
-/// Which backend a network runs on and its per-node specs.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub enum BackendSpec {
-    Native {
-        nodes: Vec<NativeNodeSpec>,
-    },
-    Docker {
-        /// Docker network to create + attach every container to.
-        network_name: String,
-        nodes: Vec<DockerNodeSpec>,
-    },
-}
-
-/// Everything the supervisor needs to run a network. Held as the in-memory SSOT
-/// for the process lifetime; never re-read from disk.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct SupervisorPlan {
-    pub network_id: String,
-    pub socket_path: PathBuf,
-    pub spec: BackendSpec,
-}
-
-// ---------------------------------------------------------------------------
-// Live state
-// ---------------------------------------------------------------------------
-
-/// Per-node status. `Exited` carries the reaped code so migrate-exit
-/// clean-vs-crash is distinguishable downstream.
-#[derive(Clone, Debug, Serialize)]
-#[serde(tag = "state", rename_all = "snake_case")]
-pub enum NodeStatus {
-    Running { pid: Option<u32> },
-    Exited { code: Option<i32> },
-    Failed { error: String },
-}
-
-struct SupervisorState {
-    nodes: HashMap<String, NodeStatus>,
-    /// Kill handles for live units, for teardown.
-    killers: HashMap<String, Killer>,
-    shutdown: bool,
-}
-
-impl SupervisorState {
-    fn new() -> Self {
-        SupervisorState {
-            nodes: HashMap::new(),
-            killers: HashMap::new(),
-            shutdown: false,
-        }
-    }
-
-    /// Snapshot for the `status` RPC (ordered by node name for stable output).
-    fn snapshot(&self) -> serde_json::Value {
-        let mut names: Vec<&String> = self.nodes.keys().collect();
-        names.sort();
-        let nodes: Vec<serde_json::Value> = names
-            .into_iter()
-            .map(|n| serde_json::json!({ "name": n, "status": self.nodes[n] }))
-            .collect();
-        serde_json::json!({ "nodes": nodes })
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Units: the backend-agnostic runnable
-// ---------------------------------------------------------------------------
-
-/// A running unit the supervisor awaits for exit. Owned by its waiter task.
-enum RunningUnit {
-    Native(tokio::process::Child),
-    Docker { docker: Docker, name: String },
-}
-
-impl RunningUnit {
-    /// Await the unit's exit and return its exit code (`None` if unknown).
-    async fn wait(&mut self) -> Option<i32> {
-        match self {
-            RunningUnit::Native(child) => match child.wait().await {
-                Ok(status) => status.code(),
-                Err(e) => {
-                    warn!("supervisor: wait() failed: {e}");
-                    None
-                }
-            },
-            RunningUnit::Docker { docker, name } => {
-                let mut stream = docker.wait_container(name, None::<WaitContainerOptions<String>>);
-                match stream.next().await {
-                    Some(Ok(r)) => Some(r.status_code as i32),
-                    Some(Err(e)) => {
-                        warn!("supervisor: wait_container '{name}' failed: {e}");
-                        None
-                    }
-                    None => None,
-                }
-            }
-        }
-    }
-}
-
-/// A cheap, cloneable handle to terminate a unit during teardown.
-#[derive(Clone)]
-enum Killer {
-    Native { pid: u32 },
-    Docker { docker: Docker, name: String },
-}
-
-impl Killer {
-    /// Graceful stop (SIGTERM / `docker stop`).
-    async fn terminate(&self) {
-        match self {
-            Killer::Native { pid } => {
-                use nix::sys::signal::{self, Signal};
-                use nix::unistd::Pid;
-                let _ = signal::kill(Pid::from_raw(*pid as i32), Signal::SIGTERM);
-            }
-            Killer::Docker { docker, name } => {
-                let _ = docker
-                    .stop_container(name, Some(StopContainerOptions { t: 2 }))
-                    .await;
-            }
-        }
-    }
-
-    /// Forceful removal (SIGKILL survivors / `docker rm -f`).
-    async fn force_kill(&self) {
-        match self {
-            Killer::Native { pid } => {
-                if process_alive(*pid) {
-                    use nix::sys::signal::{self, Signal};
-                    use nix::unistd::Pid;
-                    let _ = signal::kill(Pid::from_raw(*pid as i32), Signal::SIGKILL);
-                }
-            }
-            Killer::Docker { docker, name } => {
-                let _ = docker
-                    .remove_container(
-                        name,
-                        Some(RemoveContainerOptions {
-                            force: true,
-                            ..Default::default()
-                        }),
-                    )
-                    .await;
-            }
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Launching
-// ---------------------------------------------------------------------------
-
-/// Spawn a native daemon as an owned child. Sets `PR_SET_PDEATHSIG(SIGKILL)` so
-/// the child dies if the supervisor dies (best-effort backstop; teardown does an
-/// explicit kill).
-fn spawn_native(node: &NativeNodeSpec) -> std::io::Result<tokio::process::Child> {
-    use std::os::unix::process::CommandExt;
-
-    let log = std::fs::File::create(&node.log_file)?;
-    let log_err = log.try_clone()?;
-
-    let mut cmd = std::process::Command::new(&node.binary);
-    cmd.args(&node.args);
-    for (k, v) in &node.env {
-        cmd.env(k, v);
-    }
-    cmd.stdout(std::process::Stdio::from(log));
-    cmd.stderr(std::process::Stdio::from(log_err));
-
-    // SAFETY: `pre_exec` runs in the child after fork, before exec. `prctl` is
-    // async-signal-safe. pdeathsig is per-thread ⇒ best-effort backstop on a
-    // multi-thread runtime; explicit teardown is the real guarantee.
-    unsafe {
-        cmd.pre_exec(|| {
-            let r = nix::libc::prctl(
-                nix::libc::PR_SET_PDEATHSIG,
-                nix::libc::SIGKILL as nix::libc::c_ulong,
-            );
-            if r != 0 {
-                return Err(std::io::Error::last_os_error());
-            }
-            Ok(())
-        });
-    }
-
-    let mut tokio_cmd = tokio::process::Command::from(cmd);
-    tokio_cmd.kill_on_drop(true);
-    tokio_cmd.spawn()
-}
-
-/// Connect to the docker daemon and create the network (labelled for grouping).
-async fn docker_connect_and_network(
-    network_name: &str,
-    network_id: &str,
-) -> std::io::Result<Docker> {
-    let docker = Docker::connect_with_local_defaults()
-        .map_err(|e| std::io::Error::other(format!("docker connect failed: {e}")))?;
-    let mut labels = HashMap::new();
-    labels.insert("minimina.network".to_string(), network_id.to_string());
-    // Best-effort: ignore "already exists" so re-runs don't hard-fail.
-    if let Err(e) = docker
-        .create_network(CreateNetworkOptions {
-            name: network_name.to_string(),
-            labels,
-            ..Default::default()
-        })
-        .await
-    {
-        warn!("supervisor: create_network '{network_name}' (continuing): {e}");
-    }
-    Ok(docker)
-}
-
-/// Pull the image, create + start the container, return its running unit, kill
-/// handle, and host-side pid.
-async fn launch_docker(
-    docker: &Docker,
-    network_name: &str,
-    network_id: &str,
-    node: &DockerNodeSpec,
-) -> std::io::Result<(RunningUnit, Killer, Option<u32>)> {
-    // Pull (cached if already present).
-    let mut pull = docker.create_image(
-        Some(CreateImageOptions {
-            from_image: node.image.clone(),
-            ..Default::default()
-        }),
-        None,
-        None,
-    );
-    while let Some(item) = pull.next().await {
-        item.map_err(|e| std::io::Error::other(format!("pull '{}' failed: {e}", node.image)))?;
-    }
-
-    // Port bindings + exposed ports.
-    let mut port_bindings = HashMap::new();
-    let mut exposed = HashMap::new();
-    for (host, container) in &node.ports {
-        let key = format!("{container}/tcp");
-        port_bindings.insert(
-            key.clone(),
-            Some(vec![PortBinding {
-                host_ip: Some("0.0.0.0".to_string()),
-                host_port: Some(host.to_string()),
-            }]),
-        );
-        exposed.insert(key, HashMap::new());
-    }
-
-    let binds: Vec<String> = node
-        .mounts
-        .iter()
-        .map(|m| {
-            if m.read_only {
-                format!("{}:{}:ro", m.host, m.container)
-            } else {
-                format!("{}:{}", m.host, m.container)
-            }
-        })
-        .collect();
-
-    let mut endpoints = HashMap::new();
-    endpoints.insert(
-        network_name.to_string(),
-        EndpointSettings {
-            aliases: Some(node.aliases.clone()),
-            ..Default::default()
-        },
-    );
-
-    let mut labels = HashMap::new();
-    labels.insert("minimina.network".to_string(), network_id.to_string());
-
-    let host_config = HostConfig {
-        binds: if binds.is_empty() { None } else { Some(binds) },
-        port_bindings: if port_bindings.is_empty() {
-            None
-        } else {
-            Some(port_bindings)
-        },
-        ..Default::default()
-    };
-
-    docker
-        .create_container(
-            Some(CreateContainerOptions {
-                name: node.name.clone(),
-                platform: None,
-            }),
-            Config {
-                image: Some(node.image.clone()),
-                entrypoint: node.entrypoint.clone(),
-                cmd: if node.cmd.is_empty() {
-                    None
-                } else {
-                    Some(node.cmd.clone())
-                },
-                env: Some(node.env.iter().map(|(k, v)| format!("{k}={v}")).collect()),
-                exposed_ports: if exposed.is_empty() {
-                    None
-                } else {
-                    Some(exposed)
-                },
-                labels: Some(labels),
-                host_config: Some(host_config),
-                networking_config: Some(NetworkingConfig {
-                    endpoints_config: endpoints,
-                }),
-                ..Default::default()
-            },
-        )
-        .await
-        .map_err(|e| {
-            std::io::Error::other(format!("create_container '{}' failed: {e}", node.name))
-        })?;
-
-    docker
-        .start_container(
-            &node.name,
-            None::<bollard::container::StartContainerOptions<String>>,
-        )
-        .await
-        .map_err(|e| {
-            std::io::Error::other(format!("start_container '{}' failed: {e}", node.name))
-        })?;
-
-    let pid = docker
-        .inspect_container(&node.name, None)
-        .await
-        .ok()
-        .and_then(|i| i.state)
-        .and_then(|s| s.pid)
-        .map(|p| p as u32);
-
-    Ok((
-        RunningUnit::Docker {
-            docker: docker.clone(),
-            name: node.name.clone(),
-        },
-        Killer::Docker {
-            docker: docker.clone(),
-            name: node.name.clone(),
-        },
-        pid,
-    ))
-}
-
-// ---------------------------------------------------------------------------
-// JSON-RPC 2.0 envelope (hand-rolled)
-// ---------------------------------------------------------------------------
-
-#[derive(Deserialize)]
-struct RpcRequest {
-    #[allow(dead_code)]
-    jsonrpc: Option<String>,
-    #[serde(default)]
-    id: serde_json::Value,
-    method: String,
-    #[allow(dead_code)]
-    #[serde(default)]
-    params: serde_json::Value,
-}
-
-#[derive(Serialize)]
-struct RpcResponse {
-    jsonrpc: &'static str,
-    id: serde_json::Value,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    result: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    error: Option<RpcError>,
-}
-
-#[derive(Serialize)]
-struct RpcError {
-    code: i64,
-    message: String,
-}
-
-impl RpcResponse {
-    fn ok(id: serde_json::Value, result: serde_json::Value) -> Self {
-        RpcResponse {
-            jsonrpc: "2.0",
-            id,
-            result: Some(result),
-            error: None,
-        }
-    }
-    fn err(id: serde_json::Value, code: i64, message: String) -> Self {
-        RpcResponse {
-            jsonrpc: "2.0",
-            id,
-            result: None,
-            error: Some(RpcError { code, message }),
-        }
-    }
-}
-
-const METHOD_NOT_FOUND: i64 = -32601;
-
-// ---------------------------------------------------------------------------
-// Entry point
-// ---------------------------------------------------------------------------
-
-/// Build a tokio runtime and run the supervisor to completion (blocking).
-/// Returns when the network is torn down (via `stop` RPC or SIGINT).
-pub fn run_blocking(plan: SupervisorPlan) -> std::io::Result<()> {
-    let rt = tokio::runtime::Builder::new_multi_thread()
-        .enable_all()
-        .build()?;
-    rt.block_on(run(plan))
-}
-
-async fn run(plan: SupervisorPlan) -> std::io::Result<()> {
-    let state = Arc::new(Mutex::new(SupervisorState::new()));
-    let shutdown = Arc::new(Notify::new());
-
-    // Launch every unit + a waiter task that reaps its exit.
-    let docker = match &plan.spec {
-        BackendSpec::Native { nodes } => {
-            for node in nodes {
-                match spawn_native(node) {
-                    Ok(child) => {
-                        let pid = child.id();
-                        register_unit(
-                            &state,
-                            &node.name,
-                            pid,
-                            RunningUnit::Native(child),
-                            Killer::Native {
-                                pid: pid.unwrap_or(0),
-                            },
-                        );
-                    }
-                    Err(e) => fail_unit(&state, &node.name, e.to_string()),
-                }
-            }
-            None
-        }
-        BackendSpec::Docker {
-            network_name,
-            nodes,
-        } => {
-            let docker = docker_connect_and_network(network_name, &plan.network_id).await?;
-            for node in nodes {
-                match launch_docker(&docker, network_name, &plan.network_id, node).await {
-                    Ok((unit, killer, pid)) => register_unit(&state, &node.name, pid, unit, killer),
-                    Err(e) => fail_unit(&state, &node.name, e.to_string()),
-                }
-            }
-            Some((docker, network_name.clone()))
-        }
-    };
-
-    // Bind the RPC socket (unlink any stale socket first).
-    let _ = std::fs::remove_file(&plan.socket_path);
-    let listener = UnixListener::bind(&plan.socket_path)?;
-    info!(
-        "supervisor: serving RPC on '{}'",
-        plan.socket_path.display()
-    );
-
-    let accept_state = state.clone();
-    let accept_shutdown = shutdown.clone();
-    let accept_task = tokio::spawn(async move {
-        loop {
-            match listener.accept().await {
-                Ok((stream, _addr)) => {
-                    tokio::spawn(handle_connection(
-                        stream,
-                        accept_state.clone(),
-                        accept_shutdown.clone(),
-                    ));
-                }
-                Err(e) => {
-                    warn!("supervisor: accept error: {e}");
-                    break;
-                }
-            }
-        }
-    });
-
-    tokio::select! {
-        _ = shutdown.notified() => info!("supervisor: stop requested"),
-        r = tokio::signal::ctrl_c() => match r {
-            Ok(()) => info!("supervisor: SIGINT received"),
-            Err(e) => warn!("supervisor: signal error: {e}"),
-        },
-    }
-
-    accept_task.abort();
-    teardown(&state, &docker).await;
-    let _ = std::fs::remove_file(&plan.socket_path);
-    info!("supervisor: network '{}' stopped", plan.network_id);
-    Ok(())
-}
-
-/// Record a launched unit in state and spawn its waiter task.
-fn register_unit(
-    state: &Arc<Mutex<SupervisorState>>,
-    name: &str,
-    pid: Option<u32>,
-    mut unit: RunningUnit,
-    killer: Killer,
-) {
-    info!("supervisor: started '{name}' (pid {pid:?})");
-    {
-        let mut st = state.lock().unwrap();
-        st.nodes
-            .insert(name.to_string(), NodeStatus::Running { pid });
-        st.killers.insert(name.to_string(), killer);
-    }
-    let st = state.clone();
-    let name = name.to_string();
-    tokio::spawn(async move {
-        let code = unit.wait().await;
-        info!("supervisor: '{name}' exited (code {code:?})");
-        let mut guard = st.lock().unwrap();
-        guard
-            .nodes
-            .insert(name.clone(), NodeStatus::Exited { code });
-        guard.killers.remove(&name);
-    });
-}
-
-fn fail_unit(state: &Arc<Mutex<SupervisorState>>, name: &str, error: String) {
-    error!("supervisor: failed to start '{name}': {error}");
-    state
-        .lock()
-        .unwrap()
-        .nodes
-        .insert(name.to_string(), NodeStatus::Failed { error });
-}
-
-/// Handle one client connection: newline-delimited JSON-RPC request/response.
-async fn handle_connection(
-    stream: tokio::net::UnixStream,
-    state: Arc<Mutex<SupervisorState>>,
-    shutdown: Arc<Notify>,
-) {
-    let (reader, mut writer) = stream.into_split();
-    let mut lines = BufReader::new(reader).lines();
-    loop {
-        let line = match lines.next_line().await {
-            Ok(Some(l)) => l,
-            Ok(None) => break,
-            Err(e) => {
-                warn!("supervisor: read error: {e}");
-                break;
-            }
-        };
-        if line.trim().is_empty() {
-            continue;
-        }
-
-        let response = match serde_json::from_str::<RpcRequest>(&line) {
-            Ok(req) => dispatch(req, &state, &shutdown),
-            Err(e) => RpcResponse::err(
-                serde_json::Value::Null,
-                -32700,
-                format!("invalid request: {e}"),
-            ),
-        };
-
-        let mut buf = match serde_json::to_vec(&response) {
-            Ok(b) => b,
-            Err(e) => {
-                error!("supervisor: failed to serialize response: {e}");
-                break;
-            }
-        };
-        buf.push(b'\n');
-        if let Err(e) = writer.write_all(&buf).await {
-            warn!("supervisor: write error: {e}");
-            break;
-        }
-    }
-}
-
-/// Dispatch a single request. Locks state only briefly, never across an await.
-fn dispatch(
-    req: RpcRequest,
-    state: &Arc<Mutex<SupervisorState>>,
-    shutdown: &Arc<Notify>,
-) -> RpcResponse {
-    match req.method.as_str() {
-        "status" => {
-            let snap = state.lock().unwrap().snapshot();
-            RpcResponse::ok(req.id, snap)
-        }
-        "stop" => {
-            state.lock().unwrap().shutdown = true;
-            shutdown.notify_one();
-            RpcResponse::ok(req.id, serde_json::json!({ "stopping": true }))
-        }
-        other => RpcResponse::err(
-            req.id,
-            METHOD_NOT_FOUND,
-            format!("unknown method '{other}'"),
-        ),
-    }
-}
-
-/// Terminate every live unit (graceful → force), then remove the docker network.
-async fn teardown(state: &Arc<Mutex<SupervisorState>>, docker: &Option<(Docker, String)>) {
-    let killers: Vec<Killer> = {
-        let st = state.lock().unwrap();
-        st.killers.values().cloned().collect()
-    };
-    for k in &killers {
-        k.terminate().await;
-    }
-    if !killers.is_empty() {
-        tokio::time::sleep(std::time::Duration::from_secs(2)).await;
-        for k in &killers {
-            k.force_kill().await;
-        }
-    }
-    if let Some((docker, network_name)) = docker {
-        if let Err(e) = docker.remove_network(network_name).await {
-            warn!("supervisor: remove_network '{network_name}': {e}");
-        }
-    }
-}
-
-fn process_alive(pid: u32) -> bool {
-    use nix::sys::signal;
-    use nix::unistd::Pid;
-    signal::kill(Pid::from_raw(pid as i32), None).is_ok()
-}
-
-// ---------------------------------------------------------------------------
-// Blocking client (used by the CLI's `status`/`stop` as thin RPC clients)
-// ---------------------------------------------------------------------------
-
-/// Send one JSON-RPC request over the socket and return the `result` value.
-/// Blocking + synchronous — usable from the otherwise non-async CLI.
-pub fn rpc_call(
-    socket_path: &Path,
-    method: &str,
-    params: serde_json::Value,
-) -> std::io::Result<serde_json::Value> {
-    use std::io::{BufRead, BufReader as StdBufReader, Write};
-    use std::os::unix::net::UnixStream as StdUnixStream;
-
-    let mut stream = StdUnixStream::connect(socket_path)?;
-    let req = serde_json::json!({
-        "jsonrpc": "2.0",
-        "id": 1,
-        "method": method,
-        "params": params,
-    });
-    let mut line = serde_json::to_vec(&req)?;
-    line.push(b'\n');
-    stream.write_all(&line)?;
-    stream.flush()?;
-
-    let mut reader = StdBufReader::new(stream);
-    let mut resp_line = String::new();
-    reader.read_line(&mut resp_line)?;
-
-    let resp: serde_json::Value = serde_json::from_str(resp_line.trim())?;
-    if let Some(err) = resp.get("error") {
-        if !err.is_null() {
-            return Err(std::io::Error::other(format!("rpc error: {err}")));
-        }
-    }
-    Ok(resp
-        .get("result")
-        .cloned()
-        .unwrap_or(serde_json::Value::Null))
-}
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// End-to-end native: launch a real child, query `status`, then `stop`, and
-    /// confirm the child is reaped and the socket is cleaned up.
-    #[test]
-    fn supervise_status_stop_reaps_child() {
-        let dir = tempdir::TempDir::new("supervisor-test").unwrap();
-        let socket_path = dir.path().join("supervisor.sock");
-        let log_file = dir.path().join("sleeper.log");
-
-        let plan = SupervisorPlan {
-            network_id: "test-net".into(),
-            socket_path: socket_path.clone(),
-            spec: BackendSpec::Native {
-                nodes: vec![NativeNodeSpec {
-                    name: "sleeper".into(),
-                    binary: "/bin/sleep".into(),
-                    args: vec!["300".into()],
-                    env: vec![],
-                    log_file,
-                }],
-            },
-        };
-
-        let sup = std::thread::spawn(move || run_blocking(plan).unwrap());
-
-        let mut waited = 0;
-        while !socket_path.exists() && waited < 100 {
-            std::thread::sleep(std::time::Duration::from_millis(50));
-            waited += 1;
-        }
-        assert!(socket_path.exists(), "socket never appeared");
-
-        let status = rpc_call(&socket_path, "status", serde_json::Value::Null).unwrap();
-        let nodes = status["nodes"].as_array().unwrap();
-        assert_eq!(nodes.len(), 1);
-        assert_eq!(nodes[0]["name"], "sleeper");
-        assert_eq!(nodes[0]["status"]["state"], "running");
-        let pid = nodes[0]["status"]["pid"].as_u64().unwrap() as u32;
-        assert!(process_alive(pid), "sleeper should be alive");
-
-        let stop = rpc_call(&socket_path, "stop", serde_json::Value::Null).unwrap();
-        assert_eq!(stop["stopping"], true);
-
-        sup.join().unwrap();
-
-        assert!(!process_alive(pid), "sleeper should have been killed");
-        assert!(!socket_path.exists(), "socket should be removed");
-    }
-
-    #[test]
-    fn unknown_method_returns_error() {
-        let state = Arc::new(Mutex::new(SupervisorState::new()));
-        let shutdown = Arc::new(Notify::new());
-        let req = RpcRequest {
-            jsonrpc: Some("2.0".into()),
-            id: serde_json::json!(7),
-            method: "nope".into(),
-            params: serde_json::Value::Null,
-        };
-        let resp = dispatch(req, &state, &shutdown);
-        assert_eq!(resp.id, serde_json::json!(7));
-        assert!(resp.result.is_none());
-        assert_eq!(resp.error.unwrap().code, METHOD_NOT_FOUND);
-    }
-
-    /// End-to-end docker: launch a real alpine container as a supervisor unit,
-    /// query `status`, then `stop`, and confirm teardown removes it. Requires a
-    /// docker daemon; `#[ignore]`d so CI (no docker-in-docker) skips it.
-    /// Run manually: `cargo test supervise_docker_unit -- --ignored --nocapture`.
-    #[test]
-    #[ignore]
-    fn supervise_docker_unit_status_stop() {
-        let dir = tempdir::TempDir::new("supervisor-docker-test").unwrap();
-        let socket_path = dir.path().join("supervisor.sock");
-
-        let plan = SupervisorPlan {
-            network_id: "docker-test-net".into(),
-            socket_path: socket_path.clone(),
-            spec: BackendSpec::Docker {
-                network_name: "minimina-suptest-net".into(),
-                nodes: vec![DockerNodeSpec {
-                    name: "minimina-suptest-ctr".into(),
-                    image: "alpine:3.19".into(),
-                    entrypoint: None,
-                    cmd: vec!["sleep".into(), "300".into()],
-                    env: vec![],
-                    ports: vec![],
-                    mounts: vec![],
-                    aliases: vec!["suptest-node".into()],
-                }],
-            },
-        };
-
-        let sup = std::thread::spawn(move || run_blocking(plan).unwrap());
-
-        let mut waited = 0;
-        while !socket_path.exists() && waited < 600 {
-            std::thread::sleep(std::time::Duration::from_millis(100));
-            waited += 1;
-        }
-        assert!(socket_path.exists(), "socket never appeared");
-
-        let status = rpc_call(&socket_path, "status", serde_json::Value::Null).unwrap();
-        let nodes = status["nodes"].as_array().unwrap();
-        assert_eq!(nodes.len(), 1);
-        assert_eq!(nodes[0]["status"]["state"], "running");
-
-        let _ = rpc_call(&socket_path, "stop", serde_json::Value::Null).unwrap();
-        sup.join().unwrap();
-        assert!(!socket_path.exists(), "socket should be removed");
-    }
-}
